refactor(packs): add repository layer with context support
- Created repository.go with 21 repository functions
- All functions now accept ctx context.Context as first parameter
- Updated all function calls in packs.go to pass context
- Exported FindPackByID and CheckPackOwnership for use by images package
- Updated pkg/images to use context-aware repository functions
- Updated tests to pass context.Background()
- Removed duplicate repository function definitions from packs.go
- All tests pass (46.0% coverage maintained)
